[remove] : FCMService 와 테스트 추가 - 테스트 오류 변경 필요, BookmarkService 와 테스트 수정
